Enforce ESLinting on commit
Install and configure husky (a tool that makes it easy to hook into git commands) and lint-staged (lints staged files), so that no code with JS issues (as defined in .eslintrc.js) can be committed to the repo.

Note: this also pretty-prints all code on commit, since Prettier is configured to run when ESLint runs. Prettier rules are applied in .prettierrc.js -- see https://prettier.io/docs/en/options.html for options.